+++ includes/service-providers/mailchimp/class-newspack-newsletters-mailchimp-controller.php
@@ -1,0 +1,204 @@
+<?php
+/**
+ * Mailchimp ESP Service Controller.
+ *
+ * @package Newspack
+ */
+
+defined( 'ABSPATH' ) || exit;
+
+use \DrewM\MailChimp\MailChimp;
+
+/**
+ * API Controller for Newspack Mailchimp ESP service.
+ */
+class Newspack_Newsletters_Mailchimp_Controller extends Newspack_Newsletters_ESP_Service_Provider_Controller {
+
+	/**
+	 * @var \Newspack_Newsletters_Mailchimp
+	 */
+	private $mailchimp;
+
+	/**
+	 * Newspack_Newsletters_Mailchimp_Controller constructor.
+	 *
+	 * @param \Newspack_Newsletters_Mailchimp $mailchimp
+	 */
+	public function __construct( $mailchimp ) {
+		$this->service_provider = $mailchimp;
+		add_action( 'rest_api_init', [ $this, 'register_routes' ] );
+		parent::__construct( $mailchimp );
+	}
+
+	/**
+	 * Register API endpoints unique to Mailchimp.
+	 */
+	public function register_routes() {
+
+		// Register common ESP routes from \Newspack_Newsletters_ESP_Service_Provider_Controller::register_routes.
+		parent::register_routes();
+
+		\register_rest_route(
+			$this->service_provider::BASE_NAMESPACE . $this->service_provider->service,
+			'(?P<id>[\a-z]+)',
+			[
+				'methods'             => \WP_REST_Server::READABLE,
+				'callback'            => [ $this, 'api_retrieve' ],
+				'permission_callback' => [ $this->service_provider, 'api_authoring_permissions_check' ],
+				'args'                => [
+					'id' => [
+						'sanitize_callback' => 'absint',
+						'validate_callback' => [ 'Newspack_Newsletters', 'validate_newsletter_id' ],
+					],
+				],
+			]
+		);
+		\register_rest_route(
+			$this->service_provider::BASE_NAMESPACE . $this->service_provider->service,
+			'(?P<id>[\a-z]+)/test',
+			[
+				'methods'             => \WP_REST_Server::EDITABLE,
+				'callback'            => [ $this, 'api_test' ],
+				'permission_callback' => [ $this->service_provider, 'api_authoring_permissions_check' ],
+				'args'                => [
+					'id'         => [
+						'sanitize_callback' => 'absint',
+						'validate_callback' => [ 'Newspack_Newsletters', 'validate_newsletter_id' ],
+					],
+					'test_email' => [
+						'sanitize_callback' => 'sanitize_text_field',
+					],
+				],
+			]
+		);
+		\register_rest_route(
+			$this->service_provider::BASE_NAMESPACE . $this->service_provider->service,
+			'(?P<id>[\a-z]+)/sender',
+			[
+				'methods'             => \WP_REST_Server::EDITABLE,
+				'callback'            => [ $this, 'api_sender' ],
+				'permission_callback' => [ $this->service_provider, 'api_authoring_permissions_check' ],
+				'args'                => [
+					'id'        => [
+						'sanitize_callback' => 'absint',
+						'validate_callback' => [ 'Newspack_Newsletters', 'validate_newsletter_id' ],
+					],
+					'from_name' => [
+						'sanitize_callback' => 'sanitize_text_field',
+					],
+					'reply_to'  => [
+						'sanitize_callback' => 'sanitize_email',
+					],
+				],
+			]
+		);
+		\register_rest_route(
+			$this->service_provider::BASE_NAMESPACE . $this->service_provider->service,
+			'(?P<id>[\a-z]+)/list/(?P<list_id>[\a-z]+)',
+			[
+				'methods'             => \WP_REST_Server::EDITABLE,
+				'callback'            => [ $this, 'api_list' ],
+				'permission_callback' => [ $this->service_provider, 'api_authoring_permissions_check' ],
+				'args'                => [
+					'id'      => [
+						'sanitize_callback' => 'absint',
+						'validate_callback' => [ 'Newspack_Newsletters', 'validate_newsletter_id' ],
+					],
+					'list_id' => [
+						'sanitize_callback' => 'esc_attr',
+					],
+				],
+			]
+		);
+		\register_rest_route(
+			$this->service_provider::BASE_NAMESPACE . $this->service_provider->service,
+			'(?P<id>[\a-z]+)/interest/(?P<interest_id>[\a-z]+)',
+			[
+				'methods'             => \WP_REST_Server::EDITABLE,
+				'callback'            => [ $this, 'api_interest' ],
+				'permission_callback' => [ $this->service_provider, 'api_authoring_permissions_check' ],
+				'args'                => [
+					'id'          => [
+						'sanitize_callback' => 'absint',
+						'validate_callback' => [ 'Newspack_Newsletters', 'validate_newsletter_id' ],
+					],
+					'interest_id' => [
+						'sanitize_callback' => 'esc_attr',
+					],
+				],
+			]
+		);
+	}
+
+	/**
+	 * Get campaign data.
+	 *
+	 * @param WP_REST_Request $request API request object.
+	 * @return WP_REST_Response|mixed API response or error.
+	 */
+	public function api_retrieve( $request ) {
+		$response = $this->service_provider->retrieve( $request['id'] );
+		return \rest_ensure_response( $response );
+	}
+
+	/**
+	 * Test campaign.
+	 *
+	 * @param WP_REST_Request $request API request object.
+	 * @return WP_REST_Response|mixed API response or error.
+	 */
+	public function api_test( $request ) {
+		$emails = explode( ',', $request['test_email'] );
+		foreach ( $emails as &$email ) {
+			$email = sanitize_email( trim( $email ) );
+		}
+		$response = $this->service_provider->test(
+			$request['id'],
+			$emails
+		);
+		return \rest_ensure_response( $response );
+	}
+
+	/**
+	 * Set the sender name and email for the campaign.
+	 *
+	 * @param WP_REST_Request $request API request object.
+	 * @return WP_REST_Response|mixed API response or error.
+	 */
+	public function api_sender( $request ) {
+		$response = $this->service_provider->sender(
+			$request['id'],
+			$request['from_name'],
+			$request['reply_to']
+		);
+		return \rest_ensure_response( $response );
+	}
+
+	/**
+	 * Set list for a campaign.
+	 *
+	 * @param WP_REST_Request $request API request object.
+	 * @return WP_REST_Response|mixed API response or error.
+	 */
+	public function api_list( $request ) {
+		$response = $this->service_provider->list(
+			$request['id'],
+			$request['list_id']
+		);
+		return \rest_ensure_response( $response );
+	}
+
+	/**
+	 * Set Mailchimp interest (group) for a campaign.
+	 *
+	 * @param WP_REST_Request $request API request object.
+	 * @return WP_REST_Response|mixed API response or error.
+	 */
+	public function api_interest( $request ) {
+		$response = $this->service_provider->interest(
+			$request['id'],
+			$request['interest_id']
+		);
+		return \rest_ensure_response( $response );
+	}
+}
