+++ includes/service-providers/mailchimp/class-newspack-newsletters-mailchimp.php
@@ -12,19 +12,22 @@ use \DrewM\MailChimp\MailChimp;
 /**
  * Main Newspack Newsletters Class.
  */
-final class Newspack_Newsletters_Mailchimp extends Newspack_Newsletters_Service_Provider {
+final class Newspack_Newsletters_Mailchimp extends \Newspack_Newsletters_Service_Provider {
 
 	/**
 	 * Class constructor.
 	 */
 	public function __construct() {
-		$this->service = 'mailchimp';
-		parent::__construct();
-	}
+		$this->service    = 'mailchimp';
 
-	/**
-	 * Abstract method implementations.
-	 */
+		$this->controller = new Newspack_Newsletters_Mailchimp_Controller( $this );
+
+		add_action( 'save_post_' . Newspack_Newsletters::NEWSPACK_NEWSLETTERS_CPT, [ $this, 'save' ], 10, 3 );
+		add_action( 'publish_' . Newspack_Newsletters::NEWSPACK_NEWSLETTERS_CPT, [ $this, 'send' ], 10, 2 );
+		add_action( 'wp_trash_post', [ $this, 'trash' ], 10, 1 );
+
+		parent::__construct( $this );
+	}
 
 	/**
 	 * Get API key for service provider.
@@ -33,6 +36,26 @@ final class Newspack_Newsletters_Mailchimp extends Newspack_Newsletters_Service_
 	 */
 	public function api_key() {
 		return get_option( 'newspack_newsletters_mailchimp_api_key', false );
+	}
+
+	/**
+	 * Set the API key for the service provider.
+	 *
+	 * @param string $key API key.
+	 */
+	public function set_api_key( $key ) {
+		try {
+			$mc   = new Mailchimp( $key );
+			$ping = $mc->get( 'ping' );
+		} catch ( Exception $e ) {
+			$ping = null;
+		}
+		return $ping ?
+			update_option( 'newspack_newsletters_mailchimp_api_key', $key ) :
+			new WP_Error(
+				'newspack_newsletters_invalid_keys_mailchimp',
+				__( 'Please input a valid Mailchimp API key.', 'newspack-newsletters' )
+			);
 	}
 
 	/**
@@ -142,69 +165,6 @@ final class Newspack_Newsletters_Mailchimp extends Newspack_Newsletters_Service_
 	}
 
 	/**
-	 * Update ESP campaign after post save.
-	 *
-	 * @param string  $post_id Numeric ID of the campaign.
-	 * @param WP_Post $post The complete post object.
-	 * @param boolean $update Whether this is an existing post being updated or not.
-	 */
-	public function save( $post_id, $post, $update ) {
-		if ( ! $update ) {
-			update_post_meta( $post_id, 'template_id', -1 );
-		}
-		$status = get_post_status( $post_id );
-		if ( 'trash' === $status ) {
-			return;
-		}
-		$this->sync( $post );
-	}
-
-	/**
-	 * Send a campaign.
-	 *
-	 * @param integer $post_id Post ID to send.
-	 * @param WP_POST $post Post to send.
-	 */
-	public function send( $post_id, $post ) {
-		if ( ! Newspack_Newsletters::validate_newsletter_id( $post_id ) ) {
-			return new WP_Error(
-				'newspack_newsletters_incorrect_post_type',
-				__( 'Post is not a Newsletter.', 'newspack-newsletters' )
-			);
-		}
-
-		try {
-			$sync_result = $this->sync( $post );
-
-			if ( is_wp_error( $sync_result ) ) {
-				return $sync_result;
-			}
-
-			$mc_campaign_id = get_post_meta( $post_id, 'mc_campaign_id', true );
-			if ( ! $mc_campaign_id ) {
-				return new WP_Error(
-					'newspack_newsletters_no_campaign_id',
-					__( 'Mailchimp campaign ID not found.', 'newspack-newsletters' )
-				);
-			}
-
-			$mc = new Mailchimp( $this->api_key() );
-
-			$payload = [
-				'send_type' => 'html',
-			];
-			$result  = $this->validate(
-				$mc->post( "campaigns/$mc_campaign_id/actions/send", $payload ),
-				__( 'Error sending campaign.', 'newspack_newsletters' )
-			);
-		} catch ( Exception $e ) {
-			$transient = sprintf( 'newspack_newsletters_error_%s_%s', $post->ID, get_current_user_id() );
-			set_transient( $transient, $e->getMessage(), 45 );
-			return;
-		}
-	}
-
-	/**
 	 * Set sender data.
 	 *
 	 * @param string $post_id Numeric ID of the campaign.
@@ -247,7 +207,7 @@ final class Newspack_Newsletters_Mailchimp extends Newspack_Newsletters_Service_
 				return new WP_Error(
 					'newspack_newsletters_unverified_sender_domain',
 					sprintf(
-						// Translators: explanation that current domain is not verified, list of verified options.
+					// Translators: explanation that current domain is not verified, list of verified options.
 						__( '%1$s is not a verified domain. Verified domains for the linked Mailchimp account are: %2$s.', 'newspack-newsletters' ),
 						$domain,
 						implode( ', ', $verified_domains )
@@ -283,23 +243,49 @@ final class Newspack_Newsletters_Mailchimp extends Newspack_Newsletters_Service_
 	}
 
 	/**
-	 * Set the API key for the service provider.
+	 * Send test email or emails.
 	 *
-	 * @param string $key API key.
+	 * @param integer $post_id Numeric ID of the Newsletter post.
+	 * @param array   $emails Array of email addresses to send to.
+	 * @return object|WP_Error API Response or error.
 	 */
-	public function set_api_key( $key ) {
-		try {
-			$mc   = new Mailchimp( $key );
-			$ping = $mc->get( 'ping' );
-		} catch ( Exception $e ) {
-			$ping = null;
-		}
-		return $ping ?
-			update_option( 'newspack_newsletters_mailchimp_api_key', $key ) :
-			new WP_Error(
-				'newspack_newsletters_invalid_keys_mailchimp',
-				__( 'Please input a valid Mailchimp API key.', 'newspack-newsletters' )
+	public function test( $post_id, $emails ) {
+		$mc_campaign_id = get_post_meta( $post_id, 'mc_campaign_id', true );
+		if ( ! $mc_campaign_id ) {
+			return new WP_Error(
+				'newspack_newsletters_no_campaign_id',
+				__( 'Mailchimp campaign ID not found.', 'newspack-newsletters' )
 			);
+		}
+		try {
+			$mc      = new Mailchimp( $this->api_key() );
+			$payload = [
+				'test_emails' => $emails,
+				'send_type'   => 'html',
+			];
+			$result  = $this->validate(
+				$mc->post(
+					"campaigns/$mc_campaign_id/actions/test",
+					$payload
+				),
+				__( 'Error sending test email.', 'newspack_newsletters' )
+			);
+
+			$data            = $this->retrieve( $post_id );
+			$data['result']  = $result;
+			$data['message'] = sprintf(
+			// translators: Message after successful test email.
+				__( 'Mailchimp test sent successfully to %s.', 'newspack-newsletters' ),
+				implode( ', ', $emails )
+			);
+
+			return \rest_ensure_response( $data );
+		} catch ( Exception $e ) {
+			return new WP_Error(
+				'newspack_newsletters_mailchimp_error',
+				$e->getMessage()
+			);
+		}
 	}
 
 	/**
@@ -368,48 +354,65 @@ final class Newspack_Newsletters_Mailchimp extends Newspack_Newsletters_Service_
 	}
 
 	/**
-	 * Send test email or emails.
+	 * Update ESP campaign after post save.
 	 *
-	 * @param integer $post_id Numeric ID of the Newsletter post.
-	 * @param array   $emails Array of email addresses to send to.
-	 * @return object|WP_Error API Response or error.
+	 * @param string  $post_id Numeric ID of the campaign.
+	 * @param WP_Post $post The complete post object.
+	 * @param boolean $update Whether this is an existing post being updated or not.
 	 */
-	public function test( $post_id, $emails ) {
-		$mc_campaign_id = get_post_meta( $post_id, 'mc_campaign_id', true );
-		if ( ! $mc_campaign_id ) {
+	public function save( $post_id, $post, $update ) {
+		if ( ! $update ) {
+			update_post_meta( $post_id, 'template_id', -1 );
+		}
+		$status = get_post_status( $post_id );
+		if ( 'trash' === $status ) {
+			return;
+		}
+		$this->sync( $post );
+	}
+
+	/**
+	 * Send a campaign.
+	 *
+	 * @param integer $post_id Post ID to send.
+	 * @param WP_POST $post Post to send.
+	 */
+	public function send( $post_id, $post ) {
+		if ( ! Newspack_Newsletters::validate_newsletter_id( $post_id ) ) {
 			return new WP_Error(
-				'newspack_newsletters_no_campaign_id',
-				__( 'Mailchimp campaign ID not found.', 'newspack-newsletters' )
+				'newspack_newsletters_incorrect_post_type',
+				__( 'Post is not a Newsletter.', 'newspack-newsletters' )
 			);
 		}
+
 		try {
-			$mc      = new Mailchimp( $this->api_key() );
+			$sync_result = $this->sync( $post );
+
+			if ( is_wp_error( $sync_result ) ) {
+				return $sync_result;
+			}
+
+			$mc_campaign_id = get_post_meta( $post_id, 'mc_campaign_id', true );
+			if ( ! $mc_campaign_id ) {
+				return new WP_Error(
+					'newspack_newsletters_no_campaign_id',
+					__( 'Mailchimp campaign ID not found.', 'newspack-newsletters' )
+				);
+			}
+
+			$mc = new Mailchimp( $this->api_key() );
+
 			$payload = [
-				'test_emails' => $emails,
-				'send_type'   => 'html',
+				'send_type' => 'html',
 			];
 			$result  = $this->validate(
-				$mc->post(
-					"campaigns/$mc_campaign_id/actions/test",
-					$payload
-				),
-				__( 'Error sending test email.', 'newspack_newsletters' )
+				$mc->post( "campaigns/$mc_campaign_id/actions/send", $payload ),
+				__( 'Error sending campaign.', 'newspack_newsletters' )
 			);
-
-			$data            = $this->retrieve( $post_id );
-			$data['result']  = $result;
-			$data['message'] = sprintf(
-				// translators: Message after successful test email.
-				__( 'Mailchimp test sent successfully to %s.', 'newspack-newsletters' ),
-				implode( ', ', $emails )
-			);
-
-			return \rest_ensure_response( $data );
 		} catch ( Exception $e ) {
-			return new WP_Error(
-				'newspack_newsletters_mailchimp_error',
-				$e->getMessage()
-			);
+			$transient = sprintf( 'newspack_newsletters_error_%s_%s', $post->ID, get_current_user_id() );
+			set_transient( $transient, $e->getMessage(), 45 );
+			return;
 		}
 	}
 
@@ -444,68 +447,6 @@ final class Newspack_Newsletters_Mailchimp extends Newspack_Newsletters_Service_
 		} catch ( Exception $e ) {
 			return; // Fail silently.
 		}
-	}
-
-	/**
-	 * Register API endpoints unique to Mailchimp.
-	 */
-	public function rest_api_init() {
-		parent::rest_api_init(); // Register the required endpoints as well.
-		\register_rest_route(
-			self::BASE_NAMESPACE . $this->service,
-			'(?P<id>[\a-z]+)/list/(?P<list_id>[\a-z]+)',
-			[
-				'methods'             => \WP_REST_Server::EDITABLE,
-				'callback'            => [ $this, 'api_list' ],
-				'permission_callback' => [ $this, 'api_authoring_permissions_check' ],
-				'args'                => [
-					'id'      => [
-						'sanitize_callback' => 'absint',
-						'validate_callback' => [ 'Newspack_Newsletters', 'validate_newsletter_id' ],
-					],
-					'list_id' => [
-						'sanitize_callback' => 'esc_attr',
-					],
-				],
-			]
-		);
-		\register_rest_route(
-			self::BASE_NAMESPACE . $this->service,
-			'(?P<id>[\a-z]+)/interest/(?P<interest_id>[\a-z]+)',
-			[
-				'methods'             => \WP_REST_Server::EDITABLE,
-				'callback'            => [ $this, 'api_interest' ],
-				'permission_callback' => [ $this, 'api_authoring_permissions_check' ],
-				'args'                => [
-					'id'          => [
-						'sanitize_callback' => 'absint',
-						'validate_callback' => [ 'Newspack_Newsletters', 'validate_newsletter_id' ],
-					],
-					'interest_id' => [
-						'sanitize_callback' => 'esc_attr',
-					],
-				],
-			]
-		);
-	}
-
-	/**
-	 * API callbacks.
-	 */
-
-	/**
-	 * Set Mailchimp interest (group) for a campaign.
-	 *
-	 * @param WP_REST_Request $request API request object.
-	 * @return WP_REST_Response|mixed API response or error.
-	 */
-	public function api_interest( $request ) {
-		$response = $this->interest(
-			$request['id'],
-			$request['interest_id']
-		);
-		return \rest_ensure_response( $response );
-
 	}
 
 	/**
